changed the bar chart to a table.
