+++ tab1/plots/Balanced_Accuracy.pdf
